+++ src/vector.cpp
@@ -7,13 +7,13 @@ vector::vector(void)
 	x=y=0;
 }
 
-vector::vector(double _x, double _y)
+vector::vector(float _x, float _y)
 {
 	x=_x;
 	y=_y;
 }
 
-double operator*(vector v1, vector v2)
+float operator*(vector v1, vector v2)
 {
 	return v1.x*v2.x+v1.y*v2.y;
 }
@@ -28,18 +28,18 @@ vector operator-(vector v1, vector v2)
 	return vector(v1.x-v2.x,v1.y-v2.y);
 }
 
-vector operator*(double r, vector v2)
+vector operator*(float r, vector v2)
 {
 	return vector(r*v2.x,r*v2.y);
 }
 
-double vector::length()
+float vector::length()
 {
 	return std::sqrt(x*x+y*y);
 }
 void vector::normalize()
 {
-	double l=length();
+	float l=length();
 	x/=l;
 	y/=l;
 }
